+++ roadeye-api/roadeye-api-hub/src/test/java/org/re/test/service/TestService.java
@@ -1,4 +1,4 @@
-package org.re.common.api;
+package org.re.test.service;
 
 public interface TestService {
     void invoke();
